feat(overall): make the app work
changed back(added some APIs, changed model), finished ui

+++ rest/src/main/java/com/puter/rest/repository/SbParameterRepository.java
@@ -3,5 +3,5 @@ package com.puter.rest.repository;
 import com.puter.rest.entity.SbParameter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SbParameterRepository extends JpaRepository<SbParameter, Integer> {
+public interface SbParameterRepository extends JpaRepository<SbParameter, Long> {
 }
